Move test helpers into __init__.py
