+++ tests/test_api.py
@@ -3,7 +3,7 @@ import json
 from server.models import db, Assignment, Backup, Course, User, Version
 from server.utils import encode_id
 
-from .helpers import OkTestCase
+from tests import OkTestCase
 
 class TestAuth(OkTestCase):
     def _test_backup(self, submit):
