+++ tests/test_auth.py
@@ -1,4 +1,4 @@
-from .helpers import OkTestCase
+from tests import OkTestCase
 
 class TestAuth(OkTestCase):
     email = 'martymcfly@aol.com'
